requirements.txt
google-adk
google-cloud-discoveryengine
requests